Use Fly.io to deploy this application
The set of files we commit here is a result of running `fly launch`
master that asks a few questions.

Though, we must not deploy this particular app right away, as we do not
use database and there are not migrations (this is why we will need to
remove the corresponding to migrations block from `fly.toml` after
generation).

Another thing worth to be mentioned is this: as we haven't configured
our application to work in a multi-machines environment, it's better to
scale down number of machines to 1 (by default Fly starts 2 VMs) with
this command: `fly scale count 1`.