Make battle harder
The simplest approach we've had before might not work well for a big
group of warriors (5+) as they can easily kill the monster very quickly.

With these changes we are trying to make the battle more hard to finish:

- the Beast healing event will be happening twice as often (every 500ms)
- we adjust the healing points by an arbitrary formula: number of
  warriors / 2

Hopefully, this will make the process a bit more interesting to the
gamers.