Introduce 'Quick Join' feature
It allows users to join the Battlefield in a fast way, without entering
the required character's name in the Lobby form.

We generate player's name and strength in a special new module (which we
also use in the Battlefield), right after that we redirect user's to the
/battlefield URL with the given params.

In addition to the new /quick URL, we added a link to it from the main
(Lobby) page.