Add a few UI/UX improvements: self-aura, beast attack
With the previously added "Quick Join" feature it might be hard for
players to see themselves on a battlefield, so we highlight the player's
card.

Another small UX improvement is to let players click/tap on the Beast
image to attack it. It might be a bit harder to click on a smaller
"Attack" button below the Beast. However, it just duplicates the same
event.

A micro-UI fix is to remove right margin from player cards from the
last-child element to center them better on the screen.